move removePendingFieldsFromCustomUnit
