+++ src/libs/PolicyUtils.ts
@@ -7,7 +7,7 @@ import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
 import type {Policy, PolicyCategories, PolicyEmployeeList, PolicyTagList, PolicyTags, TaxRate} from '@src/types/onyx';
-import type {PolicyFeatureName, Rate, Tenant} from '@src/types/onyx/Policy';
+import type {PolicyFeatureName, Rate, CustomUnit, Tenant} from '@src/types/onyx/Policy';
 import type PolicyEmployee from '@src/types/onyx/PolicyEmployee';
 import type {EmptyObject} from '@src/types/utils/EmptyObject';
 import {isEmptyObject} from '@src/types/utils/EmptyObject';
@@ -477,6 +477,18 @@ const sortWorkspacesBySelected = (workspace1: WorkspaceDetails, workspace2: Work
     return workspace1.name?.toLowerCase().localeCompare(workspace2.name?.toLowerCase() ?? '') ?? 0;
 };
 
+/**
+ * Takes removes pendingFields and errorFields from a customUnit
+ */
+function removePendingFieldsFromCustomUnit(customUnit: CustomUnit): CustomUnit {
+    const cleanedCustomUnit = {...customUnit};
+
+    delete cleanedCustomUnit.pendingFields;
+    delete cleanedCustomUnit.errorFields;
+
+    return cleanedCustomUnit;
+}
+
 export {
     canEditTaxRate,
     extractPolicyIDFromPath,
@@ -531,6 +543,7 @@ export {
     getCustomUnit,
     getCustomUnitRate,
     sortWorkspacesBySelected,
+    removePendingFieldsFromCustomUnit,
 };
 
 export type {MemberEmailsToAccountIDs};
