+++ src/libs/actions/Policy/Category.ts
@@ -14,7 +14,8 @@ import ONYXKEYS from '@src/ONYXKEYS';
 import type {Policy, PolicyCategories, PolicyCategory, RecentlyUsedCategories, Report} from '@src/types/onyx';
 import type {CustomUnit} from '@src/types/onyx/Policy';
 import type {OnyxData} from '@src/types/onyx/Request';
-import {navigateWhenEnableFeature, removePendingFieldsFromCustomUnit} from './Policy';
+import {removePendingFieldsFromCustomUnit} from '@libs/PolicyUtils';
+import {navigateWhenEnableFeature} from './Policy';
 
 const allPolicies: OnyxCollection<Policy> = {};
 Onyx.connect({
